Add VIS vehicle signal support
Use the new VehicleSignals API from libqtappfw to replace the
previous signal-composer usage.  Additionally, the default units
for the vehicle speed and odometer have been switched to kilometers
with switching driven by the appropriate VSS schema value.  The
units in the translated text labels have been left alone for now,
if the Qt demo lifetime is extended this may be revisited.

Bug-AGL: SPEC-4409

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ie04dc429b02adb7d763d2e5a3bfea0a5a1932b92